Fix wrong dcell_t type for 32-bit

+++ cell.h
@@ -13,8 +13,8 @@
 	typedef int64_t cell_t;
 	typedef uint64_t ucell_t;
 #elif CELL32
-	typedef int32_t dcell_t;
-	typedef uint32_t udcell_t;
+	typedef int64_t dcell_t;
+	typedef uint64_t udcell_t;
 	typedef int32_t cell_t;
 	typedef uint32_t ucell_t;
 #elif CELL16
